# [Lagrangian.Correction] LinearSolverConstraintCorrection: Trivial optimisations for MSVC