--- a/Sofa/framework/DefaultType/src/sofa/defaulttype/SolidTypes.h
+++ b/Sofa/framework/DefaultType/src/sofa/defaulttype/SolidTypes.h
@@ -69,10 +69,11 @@ public:
     class SOFA_DEFAULTTYPE_API SpatialVector
     {
     public:
-        Vec lineVec;
-        Vec freeVec;
+        Vec lineVec{ type::NOINIT };
+        Vec freeVec{ type::NOINIT };
+
         void clear();
-        SpatialVector();
+        SpatialVector() = default;
         /**
         \param l The line vector: angular velocity, or force
         \param f The free vector: linear velocity, or torque