It seems msvc is quite (much!) less clever than gcc.

2 optimizations (only useful when not building in Release mode)
 - add and use a NOINIT RigidDeriv constructor (avoiding to init a deriv which is filled afterwards)
 - default ctor directly call the identity of its components (which is their default ctor by the way)

and 1 ridiculous optimization
 - avoid calling getPositionIntegrationFactor() in the nested loop
 
This optimization brings a whooping 30% speedup on my setup (VC2022, Release/LTO enabled)
Test scene: 3instruments_collis.scn from BeamAdapter, 2000 iterations in batch mode
 - before: 16.5669 s ( 120.723 FPS)
 - after: 12.829 s ( 155.897 FPS)

GCC was doing the same by the way with or without these optims (and is faster than msvc, even in a VM 🤔)

NOTE: do not forget to enable the LTO option for Windows when deploying, it brings full inlining & stuff and it is really useful for perf
(w/o LTO, before the optims it is 21.9883 s ( 90.9575 FPS) )

______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).