@alxbilger 2022-07-19
Ah ok, because Turner made two recent videos about LTO